+++ test/req/steps_test.exs
@@ -699,6 +699,47 @@ defmodule Req.StepsTest do
   end
 
   @tag :capture_log
+  test "retry: retry-after", c do
+    {:ok, _} = Agent.start_link(fn -> [] end, name: :timestamps)
+
+    Bypass.expect(c.bypass, "GET", "/", fn conn ->
+      now = DateTime.utc_now()
+
+      case Agent.get(:timestamps, & &1) do
+        [] ->
+          Agent.update(:timestamps, fn times -> [now | times] end)
+
+          conn
+          |> Plug.Conn.put_resp_header("retry-after", "1")
+          |> Plug.Conn.send_resp(429, "later")
+
+        [_] ->
+          target_time =
+            now
+            |> DateTime.add(1)
+            |> Req.Steps.format_http_datetime()
+
+          Agent.update(:timestamps, fn times -> [now | times] end)
+
+          conn
+          |> Plug.Conn.put_resp_header("retry-after", target_time)
+          |> Plug.Conn.send_resp(429, "later")
+
+        _ ->
+          Agent.update(:timestamps, fn times -> [now | times] end)
+          Plug.Conn.send_resp(conn, 200, "ok")
+      end
+    end)
+
+    request = Req.new(url: c.url, retry_delay: 10)
+    assert Req.get!(request).body == "ok"
+
+    [third, second, first] = Agent.get(:timestamps, & &1)
+    assert_in_delta DateTime.diff(second, first), 1, 0.01
+    assert_in_delta DateTime.diff(third, second), 1, 0.01
+  end
+
+  @tag :capture_log
   test "retry: always failing", c do
     pid = self()